+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.0.11-beta2
+# 0.0.11-beta3
 
 - Make all built-in types lowercased, following TypeScript syntax more closely
   instead of GraphQLs: